AddOn - Optimized + reduced net import disk size
