+++ docs/docs/auto-docs/components/AddOn/AddOn/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(`__namedParameters`): `JSX.Element`
 
-Defined in: [src/components/AddOn/AddOn.tsx:22](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/AddOn.tsx#L22)
+Defined in: [src/components/AddOn/AddOn.tsx:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/AddOn.tsx#L17)
 
 The AddOn component is used to wrap children within a plugin container.
 It also accepts additional properties (`extras` and `name`) to allow for
@@ -16,7 +16,7 @@ extensibility and custom naming.
 
 ### \_\_namedParameters
 
-`InterfaceAddOnProps`
+[`InterfaceAddOnProps`](../../../../types/AddOn/interface/interfaces/InterfaceAddOnProps.md)
 
 ## Returns
