+++ docs/docs/auto-docs/components/AddOn/core/AddOnEntry/AddOnEntry/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(`props`): `JSX.Element`
 
-Defined in: [src/components/AddOn/core/AddOnEntry/AddOnEntry.tsx:63](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/core/AddOnEntry/AddOnEntry.tsx#L63)
+Defined in: [src/components/AddOn/core/AddOnEntry/AddOnEntry.tsx:49](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/core/AddOnEntry/AddOnEntry.tsx#L49)
 
 A React component that represents an add-on entry, displaying its details and allowing installation or uninstallation.
 
@@ -14,7 +14,7 @@ A React component that represents an add-on entry, displaying its details and al
 
 ### props
 
-`InterfaceAddOnEntryProps`
+[`InterfaceAddOnEntryProps`](../../../../../../types/AddOn/interface/interfaces/InterfaceAddOnEntryProps.md)
 
 The properties for the component.
