+++ docs/docs/auto-docs/components/AddOn/mocks/functions/wait.md
@@ -1,0 +1,19 @@
+[Admin Docs](/)
+
+***
+
+# Function: wait()
+
+> **wait**(`ms`): `Promise`\<`void`\>
+
+Defined in: [src/components/AddOn/mocks.ts:31](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/mocks.ts#L31)
+
+## Parameters
+
+### ms
+
+`number` = `100`
+
+## Returns
+
+`Promise`\<`void`\>
