+++ docs/docs/auto-docs/components/AddOn/mocks/variables/ORGANIZATIONS_LIST_MOCK.md
@@ -1,0 +1,39 @@
+[Admin Docs](/)
+
+***
+
+# Variable: ORGANIZATIONS\_LIST\_MOCK
+
+> `const` **ORGANIZATIONS\_LIST\_MOCK**: `object`
+
+Defined in: [src/components/AddOn/mocks.ts:113](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/mocks.ts#L113)
+
+## Type declaration
+
+### request
+
+> **request**: `object`
+
+#### request.query
+
+> **query**: `DocumentNode` = `ORGANIZATIONS_LIST`
+
+#### request.variables
+
+> **variables**: `object`
+
+#### request.variables.id
+
+> **id**: `string` = `'undefined'`
+
+### result
+
+> **result**: `object`
+
+#### result.data
+
+> **data**: `object`
+
+#### result.data.organizations
+
+> **organizations**: `object`[]
