+++ docs/docs/auto-docs/components/AddOn/mocks/variables/PLUGIN_GET_MOCK.md
@@ -1,0 +1,35 @@
+[Admin Docs](/)
+
+***
+
+# Variable: PLUGIN\_GET\_MOCK
+
+> `const` **PLUGIN\_GET\_MOCK**: `object`
+
+Defined in: [src/components/AddOn/mocks.ts:67](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/mocks.ts#L67)
+
+## Type declaration
+
+### request
+
+> **request**: `object`
+
+#### request.query
+
+> **query**: `DocumentNode` = `PLUGIN_GET`
+
+### result
+
+> **result**: `object`
+
+#### result.data
+
+> **data**: `object`
+
+#### result.data.getPlugins
+
+> **getPlugins**: `object`[]
+
+#### result.loading
+
+> **loading**: `boolean` = `false`
