+++ docs/docs/auto-docs/components/AddOn/mocks/variables/client.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: client
+
+> `const` **client**: `ApolloClient`\<`NormalizedCacheObject`\>
+
+Defined in: [src/components/AddOn/mocks.ts:26](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/mocks.ts#L26)
