+++ docs/docs/auto-docs/components/AddOn/mocks/variables/httpLink.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: httpLink
+
+> `const` **httpLink**: `HttpLink`
+
+Defined in: [src/components/AddOn/mocks.ts:19](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/mocks.ts#L19)
