+++ docs/docs/auto-docs/components/AddOn/mocks/variables/pluginData.md
@@ -1,0 +1,23 @@
+[Admin Docs](/)
+
+***
+
+# Variable: pluginData
+
+> `const` **pluginData**: `object`
+
+Defined in: [src/components/AddOn/mocks.ts:47](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/mocks.ts#L47)
+
+## Type declaration
+
+### pluginCreatedBy
+
+> **pluginCreatedBy**: `string` = `'Test Creator'`
+
+### pluginDesc
+
+> **pluginDesc**: `string` = `'Test Description'`
+
+### pluginName
+
+> **pluginName**: `string` = `'Test Plugin'`
