+++ docs/docs/auto-docs/components/AddOn/mocks/variables/updatePluginStatus.md
@@ -1,0 +1,31 @@
+[Admin Docs](/)
+
+***
+
+# Variable: updatePluginStatus
+
+> `const` **updatePluginStatus**: `object`
+
+Defined in: [src/components/AddOn/mocks.ts:39](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/AddOn/mocks.ts#L39)
+
+## Type declaration
+
+### \_id
+
+> **\_id**: `string` = `'123'`
+
+### pluginCreatedBy
+
+> **pluginCreatedBy**: `string` = `'John Doe'`
+
+### pluginDesc
+
+> **pluginDesc**: `string` = `'This is a sample plugin description.'`
+
+### pluginName
+
+> **pluginName**: `string` = `'Sample Plugin'`
+
+### uninstalledOrgs
+
+> **uninstalledOrgs**: `any`[] = `[]`
