+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfaceActionProps.md
@@ -1,0 +1,27 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceActionProps
+
+Defined in: [src/types/AddOn/interface.ts:52](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L52)
+
+## Properties
+
+### children
+
+> **children**: `ReactNode`
+
+Defined in: [src/types/AddOn/interface.ts:56](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L56)
+
+The child elements to be rendered inside the action component.
+
+***
+
+### label
+
+> **label**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:61](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L61)
+
+The label to be displayed above the child elements.
