+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfaceAddOnEntryProps.md
@@ -1,0 +1,83 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceAddOnEntryProps
+
+Defined in: [src/types/AddOn/interface.ts:9](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L9)
+
+## Properties
+
+### component?
+
+> `optional` **component**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:15](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L15)
+
+***
+
+### createdBy
+
+> **createdBy**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:14](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L14)
+
+***
+
+### description?
+
+> `optional` **description**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L13)
+
+***
+
+### enabled?
+
+> `optional` **enabled**: `boolean`
+
+Defined in: [src/types/AddOn/interface.ts:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L11)
+
+***
+
+### getInstalledPlugins()
+
+> **getInstalledPlugins**: () => `void`
+
+Defined in: [src/types/AddOn/interface.ts:18](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L18)
+
+#### Returns
+
+`void`
+
+***
+
+### id
+
+> **id**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L10)
+
+***
+
+### modified?
+
+> `optional` **modified**: `boolean`
+
+Defined in: [src/types/AddOn/interface.ts:16](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L16)
+
+***
+
+### title?
+
+> `optional` **title**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L12)
+
+***
+
+### uninstalledOrgs
+
+> **uninstalledOrgs**: `string`[]
+
+Defined in: [src/types/AddOn/interface.ts:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L17)
