+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfaceAddOnProps.md
@@ -1,0 +1,31 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceAddOnProps
+
+Defined in: [src/types/AddOn/interface.ts:3](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L3)
+
+## Properties
+
+### children?
+
+> `optional` **children**: `ReactNode`
+
+Defined in: [src/types/AddOn/interface.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L6)
+
+***
+
+### extras?
+
+> `optional` **extras**: `Record`\<`string`, `unknown`\>
+
+Defined in: [src/types/AddOn/interface.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L4)
+
+***
+
+### name?
+
+> `optional` **name**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L5)
