+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfaceAddOnRegisterProps.md
@@ -1,0 +1,15 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceAddOnRegisterProps
+
+Defined in: [src/types/AddOn/interface.ts:29](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L29)
+
+## Properties
+
+### createdBy?
+
+> `optional` **createdBy**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:30](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L30)
