+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfaceFormStateTypes.md
@@ -1,0 +1,47 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceFormStateTypes
+
+Defined in: [src/types/AddOn/interface.ts:21](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L21)
+
+## Properties
+
+### installedOrgs
+
+> **installedOrgs**: \[\] \| \[`string`\]
+
+Defined in: [src/types/AddOn/interface.ts:26](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L26)
+
+***
+
+### pluginCreatedBy
+
+> **pluginCreatedBy**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:23](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L23)
+
+***
+
+### pluginDesc
+
+> **pluginDesc**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:24](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L24)
+
+***
+
+### pluginInstallStatus
+
+> **pluginInstallStatus**: `boolean`
+
+Defined in: [src/types/AddOn/interface.ts:25](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L25)
+
+***
+
+### pluginName
+
+> **pluginName**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:22](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L22)
