+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfaceMainContentProps.md
@@ -1,0 +1,17 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceMainContentProps
+
+Defined in: [src/types/AddOn/interface.ts:64](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L64)
+
+## Properties
+
+### children
+
+> **children**: `ReactNode`
+
+Defined in: [src/types/AddOn/interface.ts:68](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L68)
+
+The child elements to be rendered inside the main content container.
