+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfacePlugin.md
@@ -1,0 +1,31 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfacePlugin
+
+Defined in: [src/types/AddOn/interface.ts:46](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L46)
+
+## Properties
+
+### component
+
+> **component**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:49](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L49)
+
+***
+
+### enabled
+
+> **enabled**: `boolean`
+
+Defined in: [src/types/AddOn/interface.ts:47](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L47)
+
+***
+
+### pluginName
+
+> **pluginName**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:48](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L48)
