+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfacePluginHelper.md
@@ -1,0 +1,87 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfacePluginHelper
+
+Defined in: [src/types/AddOn/interface.ts:33](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L33)
+
+## Properties
+
+### \_id
+
+> **\_id**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:34](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L34)
+
+***
+
+### component
+
+> **component**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:43](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L43)
+
+***
+
+### enabled
+
+> **enabled**: `boolean`
+
+Defined in: [src/types/AddOn/interface.ts:41](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L41)
+
+***
+
+### installed
+
+> **installed**: `boolean`
+
+Defined in: [src/types/AddOn/interface.ts:40](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L40)
+
+***
+
+### name
+
+> **name**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:42](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L42)
+
+***
+
+### pluginCreatedBy
+
+> **pluginCreatedBy**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:37](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L37)
+
+***
+
+### pluginDesc?
+
+> `optional` **pluginDesc**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:36](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L36)
+
+***
+
+### pluginInstallStatus?
+
+> `optional` **pluginInstallStatus**: `boolean`
+
+Defined in: [src/types/AddOn/interface.ts:38](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L38)
+
+***
+
+### pluginName?
+
+> `optional` **pluginName**: `string`
+
+Defined in: [src/types/AddOn/interface.ts:35](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L35)
+
+***
+
+### uninstalledOrgs
+
+> **uninstalledOrgs**: `string`[]
+
+Defined in: [src/types/AddOn/interface.ts:39](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L39)
