+++ docs/docs/auto-docs/types/AddOn/interface/interfaces/InterfaceSidePanelProps.md
@@ -1,0 +1,27 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceSidePanelProps
+
+Defined in: [src/types/AddOn/interface.ts:71](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L71)
+
+## Properties
+
+### children
+
+> **children**: `ReactNode`
+
+Defined in: [src/types/AddOn/interface.ts:80](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L80)
+
+The child elements to be rendered inside the side panel.
+
+***
+
+### collapse?
+
+> `optional` **collapse**: `boolean`
+
+Defined in: [src/types/AddOn/interface.ts:75](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AddOn/interface.ts#L75)
+
+Whether the side panel should be collapsed.
