+++ docs/docs/auto-docs/types/education/type-aliases/EmploymentStatus.md
@@ -6,4 +6,4 @@
 
 > **EmploymentStatus**: *typeof* [`EmploymentStatus`](../variables/EmploymentStatus.md)\[keyof *typeof* [`EmploymentStatus`](../variables/EmploymentStatus.md)\]
 
-Defined in: [src/types/education.ts:21](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/education.ts#L21)
+Defined in: [src/types/education.ts:22](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/education.ts#L22)
