+++ docs/docs/auto-docs/types/education/variables/EmploymentStatus.md
@@ -6,7 +6,7 @@
 
 > `const` **EmploymentStatus**: `object`
 
-Defined in: [src/types/education.ts:21](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/education.ts#L21)
+Defined in: [src/types/education.ts:22](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/education.ts#L22)
 
 ## Type declaration
