+++ src/components/AddOn/core/AddOnRegister/AddOnRegister.spec.tsx
@@ -2,78 +2,16 @@ import React, { act } from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { MockedProvider } from '@apollo/react-testing';
-import { ADD_PLUGIN_MUTATION } from 'GraphQl/Mutations/mutations';
 import AddOnRegister from './AddOnRegister';
-import {
-  ApolloClient,
-  ApolloProvider,
-  InMemoryCache,
-  ApolloLink,
-  HttpLink,
-} from '@apollo/client';
-import type { NormalizedCacheObject } from '@apollo/client';
+import { ApolloProvider } from '@apollo/client';
 import { Provider } from 'react-redux';
 import { store } from 'state/store';
 import { BrowserRouter } from 'react-router-dom';
-import { BACKEND_URL } from 'Constant/constant';
 import i18nForTest from 'utils/i18nForTest';
 import { I18nextProvider } from 'react-i18next';
 import { toast } from 'react-toastify';
-import useLocalStorage from 'utils/useLocalstorage';
 import { vi } from 'vitest';
-
-const { getItem } = useLocalStorage();
-
-const httpLink = new HttpLink({
-  uri: BACKEND_URL,
-  headers: {
-    authorization: 'Bearer ' + getItem('token') || '',
-  },
-});
-
-const client: ApolloClient<NormalizedCacheObject> = new ApolloClient({
-  cache: new InMemoryCache(),
-  link: ApolloLink.from([httpLink]),
-});
-
-const mocks = [
-  {
-    request: {
-      query: ADD_PLUGIN_MUTATION,
-      variables: {
-        pluginName: 'Test Plugin',
-        pluginCreatedBy: 'AdminTest Creator',
-        pluginDesc: 'Test Description',
-        pluginInstallStatus: false,
-        installedOrgs: ['id'],
-      },
-    },
-    result: {
-      data: {
-        createPlugin: {
-          _id: '1',
-          pluginName: 'Test Plugin',
-          pluginCreatedBy: 'AdminTest Creator',
-          pluginDesc: 'Test Description',
-        },
-      },
-    },
-  },
-];
-
-async function wait(ms = 100): Promise<void> {
-  await act(() => {
-    return new Promise((resolve) => {
-      setTimeout(resolve, ms);
-    });
-  });
-}
-
-const pluginData = {
-  pluginName: 'Test Plugin',
-  pluginCreatedBy: 'Test Creator',
-  pluginDesc: 'Test Description',
-};
+import { client, mocks, wait, pluginData } from 'components/AddOn/mocks';
 
 vi.mock('react-toastify', () => ({
   toast: {
